CHROMIUM: Mali: rockchip: add platform/rk to kbase
This patch adds basic platform support for the ARM Mali T764 GPU in the
Rockchip RK3288 SoC.

With this patch, the GPU operates at the default frequency specified in
the attached-clock-rates node of its device tree node.

GPU DVFS and sysfs/debugfs nodes for controlling and debugging the driver
will be added in future patches.

BUG=chrome-os-partner:31652
TEST=emerge-veyron_pinky chromeos-kernel-3_14
TEST=on device: mali probes without error in /var/log/message
TEST=boot to UI with DRM patch set

Signed-off-by: Simon Xue <xxm@rock-chips.com>
Signed-off-by: Daniel Kurtz <djkurtz@chromium.org>
Change-Id: Ifbaa8b763484a9519b3799fc4c57f272eb8d320f
Reviewed-on: https://chromium-review.googlesource.com/214637
Reviewed-by: Dominik Behr <dbehr@chromium.org>